Compilation on gcc x86-64.

+++ components/media/geometry/include/media/geometry/defs.h
@@ -14,11 +14,11 @@ namespace geometry
 ///////////////////////////////////////////////////////////////////////////////////////////////////
 struct VertexWeight
 {
-  size_t joint_index;  //индекс соединения
-  float  joint_weight; //вес соединения
+  uint32_t joint_index;  //индекс соединения
+  float    joint_weight; //вес соединения
   
   VertexWeight ();
-  VertexWeight (size_t joint_index, float joint_weight);
+  VertexWeight (uint32_t joint_index, float joint_weight);
 };
 
 ///////////////////////////////////////////////////////////////////////////////////////////////////
@@ -26,10 +26,10 @@ struct VertexWeight
 ///////////////////////////////////////////////////////////////////////////////////////////////////
 struct VertexInfluence
 {
-  size_t first_weight, weights_count;
+  uint32_t first_weight, weights_count;
   
   VertexInfluence ();
-  VertexInfluence (size_t first_weight, size_t weights_count);
+  VertexInfluence (uint32_t first_weight, uint32_t weights_count);
 };
 
 #include <media/geometry/detail/defs.inl>
