+++ components/media/geometry/include/media/geometry/detail/defs.inl
@@ -6,7 +6,7 @@ inline VertexWeight::VertexWeight ()
   : joint_index (0), joint_weight (0.0f)
   {}
 
-inline VertexWeight::VertexWeight (size_t in_joint_index, float in_joint_weight)
+inline VertexWeight::VertexWeight (uint32_t in_joint_index, float in_joint_weight)
   : joint_index (in_joint_index), joint_weight (in_joint_weight)
   {}
 
@@ -18,6 +18,6 @@ inline VertexInfluence::VertexInfluence ()
   : first_weight (0), weights_count (0)
   {}
 
-inline VertexInfluence::VertexInfluence (size_t in_first_weight, size_t in_weights_count)
+inline VertexInfluence::VertexInfluence (uint32_t in_first_weight, uint32_t in_weights_count)
   : first_weight (in_first_weight), weights_count (in_weights_count)
   {}
