+++ components/media/geometry/include/media/geometry/detail/vertex_stream.inl
@@ -6,7 +6,7 @@ inline VertexDeclaration::VertexDeclaration (const VertexFormat& in_format)
   : format (&in_format), vertex_size (in_format.GetMinimalVertexSize ())
   {}
 
-inline VertexDeclaration::VertexDeclaration (const VertexFormat& in_format, size_t in_vertex_size)
+inline VertexDeclaration::VertexDeclaration (const VertexFormat& in_format, uint32_t in_vertex_size)
   : format (&in_format), vertex_size (in_vertex_size)
 {
   if (vertex_size < in_format.GetMinimalVertexSize ())
@@ -18,7 +18,7 @@ inline const VertexFormat& VertexDeclaration::Format () const
   return *format;
 }
 
-inline size_t VertexDeclaration::VertexSize () const
+inline uint32_t VertexDeclaration::VertexSize () const
 {
   return vertex_size;
 }
