+++ components/media/geometry/sources/binmesh/geometry_binmesh_loader.cpp
@@ -4,23 +4,23 @@ using namespace media::geometry;
 using namespace common;
 
 /*
- * array <> format: size_t elements_count; elements
+ * array <> format: uint32_t elements_count; elements
  *
  * File format: int header; int version; data
  *
  * data format: array<vertex_stream>; array<vertex_weight_stream>; array<vertex_buffer_desc>; array<index_buffer>; array<mesh>
  *
- * vertex_stream format: size_t vertices_count; size_t vertex_size; array<vertex_stream_channel>; binary<data>
- * vertex_stream_channel format: VertexAttributeSemantic semantic; VertexAttributeType type; size_t offset
+ * vertex_stream format: uint32_t vertices_count; uint32_t vertex_size; array<vertex_stream_channel>; binary<data>
+ * vertex_stream_channel format: VertexAttributeSemantic semantic; VertexAttributeType type; uint32_t offset
  * vertex_weight_stream format: array<VertexWeight>
  *
- * vertex_buffer_desc format: array<size_t> vertex_weigth_stream_id (0-1); array<size_t> vertex_stream_id;
+ * vertex_buffer_desc format: array<uint32_t> vertex_weigth_stream_id (0-1); array<uint32_t> vertex_stream_id;
  *
  * index_buffer format: IndexType type, array<index_type> indices
  *
- * mesh format: array<char> name; array<size_t> index_buffer (0-1); array<size_t> vertex_buffers; array<primitive> primitives;
+ * mesh format: array<char> name; array<uint32_t> index_buffer (0-1); array<uint32_t> vertex_buffers; array<primitive> primitives;
  *
- * primitive format: PrimitiveType type; array<char> material; size_t vertex_buffer; size_t first; size_t count
+ * primitive format: PrimitiveType type; array<char> material; uint32_t vertex_buffer; uint32_t first; uint32_t count
  */
 
 namespace components
@@ -70,7 +70,7 @@ class BinMeshLibraryLoader
     {
       VertexAttributeSemantic semantic;
       VertexAttributeType     type;
-      size_t                  offset;
+      uint32_t                offset;
 
       file_read (input_file, &semantic, sizeof (semantic));
       file_read (input_file, &type,     sizeof (type));
@@ -82,11 +82,11 @@ class BinMeshLibraryLoader
     }
 
       //чтение вершинного потока
-    void ReadVertexStream (size_t id)
+    void ReadVertexStream (uint32_t id)
     {
         //определение числа вершин и размера вершины
 
-      size_t vertices_count, vertex_size;
+      uint32_t vertices_count, vertex_size;
 
       file_read (input_file, &vertices_count, sizeof (vertices_count));
       file_read (input_file, &vertex_size, sizeof (vertex_size));
@@ -94,11 +94,11 @@ class BinMeshLibraryLoader
         //разбор вершинного формата
 
       VertexFormat vertex_format;
-      size_t       vertex_stream_channels_count;
+      uint32_t     vertex_stream_channels_count;
 
       file_read (input_file, &vertex_stream_channels_count, sizeof (vertex_stream_channels_count));
 
-      for (size_t i = 0; i < vertex_stream_channels_count; i++)
+      for (uint32_t i = 0; i < vertex_stream_channels_count; i++)
         ReadVertexStreamChannel (vertex_format);
 
         //создание вершинного потока
@@ -117,18 +117,18 @@ class BinMeshLibraryLoader
       //чтение вершинных потоков
     void ReadVertexStreams ()
     {
-      size_t vertex_streams_count;
+      uint32_t vertex_streams_count;
 
       file_read (input_file, &vertex_streams_count, sizeof (vertex_streams_count));
 
-      for (size_t i = 0; i < vertex_streams_count; i++)
+      for (uint32_t i = 0; i < vertex_streams_count; i++)
         ReadVertexStream (i);
     }
 
       //чтение потока вершинных весов
-    void ReadVertexWeightStream (size_t id)
+    void ReadVertexWeightStream (uint32_t id)
     {
-      size_t weights_count;
+      uint32_t weights_count;
 
       file_read (input_file, &weights_count, sizeof (weights_count));
 
@@ -148,16 +148,16 @@ class BinMeshLibraryLoader
     //чтение потоков вершинных весов
     void ReadVertexWeights ()
     {
-      size_t vertex_weights_streams_count;
+      uint32_t vertex_weights_streams_count;
 
       file_read (input_file, &vertex_weights_streams_count, sizeof (vertex_weights_streams_count));
 
-      for (size_t i = 0; i < vertex_weights_streams_count; i++)
+      for (uint32_t i = 0; i < vertex_weights_streams_count; i++)
         ReadVertexWeightStream (i);
     }
 
       //чтение вершинного буфера
-    void ReadVertexBuffer (size_t id)
+    void ReadVertexBuffer (uint32_t id)
     {
       static const char* METHOD_NAME = "media::geometry::BinMeshLibraryLoader::ReadVertexBuffer";
 
@@ -167,7 +167,7 @@ class BinMeshLibraryLoader
 
          //разбор потока вершинных весов
 
-      size_t has_weights;
+      uint32_t has_weights;
 
       file_read (input_file, &has_weights, sizeof (has_weights));
 
@@ -176,34 +176,34 @@ class BinMeshLibraryLoader
         if (has_weights > 1)
           throw xtl::format_operation_exception (METHOD_NAME, "Vertex buffer weights count more than 1");
 
-        size_t weights_id;
+        uint32_t weights_id;
 
         file_read (input_file, &weights_id, sizeof (weights_id));
 
         VertexWeightStreamMap::iterator weights_iter = vertex_weights.find (weights_id);
 
         if (weights_iter == vertex_weights.end ())
-          throw xtl::make_argument_exception (METHOD_NAME, "weights", weights_id, "Weights stream not found");
+          throw xtl::make_argument_exception (METHOD_NAME, "weights", (size_t)weights_id, "Weights stream not found");
 
         vb.AttachWeights (weights_iter->second);
       }
 
          //разбор вершинных потоков
 
-      size_t streams_count;
+      uint32_t streams_count;
 
       file_read (input_file, &streams_count, sizeof (streams_count));
 
-      for (size_t i = 0; i < streams_count; i++)
+      for (uint32_t i = 0; i < streams_count; i++)
       {
-        size_t stream_id;
+        uint32_t stream_id;
 
         file_read (input_file, &stream_id, sizeof (stream_id));
 
         VertexStreamMap::iterator stream_iter = vertex_streams.find (stream_id);
 
         if (stream_iter == vertex_streams.end ())
-          throw xtl::make_argument_exception (METHOD_NAME, "streams[i]", stream_id, "Vertex stream not found");
+          throw xtl::make_argument_exception (METHOD_NAME, "streams[i]", (size_t)stream_id, "Vertex stream not found");
 
         vb.Attach (stream_iter->second);
       }
@@ -216,22 +216,22 @@ class BinMeshLibraryLoader
       //чтение вершинных буферов
     void ReadVertexBuffers ()
     {
-      size_t vertex_buffers_count;
+      uint32_t vertex_buffers_count;
 
       file_read (input_file, &vertex_buffers_count, sizeof (vertex_buffers_count));
 
-      for (size_t i = 0; i < vertex_buffers_count; i++)
+      for (uint32_t i = 0; i < vertex_buffers_count; i++)
         ReadVertexBuffer (i);
     }
 
       //чтение индексного буфера
-    void ReadIndexBuffer (size_t id)
+    void ReadIndexBuffer (uint32_t id)
     {
       IndexType data_type = IndexType_Default;
 
       file_read (input_file, (int*)&data_type, sizeof (data_type));
 
-      size_t indices_count;
+      uint32_t indices_count;
 
       file_read (input_file, &indices_count, sizeof (indices_count));
       
@@ -251,11 +251,11 @@ class BinMeshLibraryLoader
       //чтение индексных буферов
     void ReadIndexBuffers ()
     {
-      size_t index_buffers_count;
+      uint32_t index_buffers_count;
 
       file_read (input_file, &index_buffers_count, sizeof (index_buffers_count));
 
-      for (size_t i = 0; i < index_buffers_count; i++)
+      for (uint32_t i = 0; i < index_buffers_count; i++)
         ReadIndexBuffer (i);
     }
 
@@ -268,7 +268,7 @@ class BinMeshLibraryLoader
 
       file_read (input_file, &type, sizeof (type));
 
-      size_t material_name_length;
+      uint32_t material_name_length;
 
       file_read (input_file, &material_name_length, sizeof (material_name_length));
 
@@ -278,7 +278,7 @@ class BinMeshLibraryLoader
 
       material.data ()[material_name_length] = '\0';
 
-      size_t vertex_buffer_index = 0, first = 0, count = 0, base_vertex = 0;
+      uint32_t vertex_buffer_index = 0, first = 0, count = 0, base_vertex = 0;
 
       file_read (input_file, &vertex_buffer_index, sizeof (vertex_buffer_index));
       file_read (input_file, &first,               sizeof (first));
@@ -289,7 +289,7 @@ class BinMeshLibraryLoader
         throw xtl::make_range_exception (METHOD_NAME, "vertex_buffer", vertex_buffer_index, mesh.VertexBuffersCount ());
 
       const VertexBuffer& vertex_buffer  = mesh.VertexBuffer (vertex_buffer_index);
-      size_t              vertices_count = vertex_buffer.VerticesCount (),
+      uint32_t            vertices_count = vertex_buffer.VerticesCount (),
                           indices_count  = mesh.IndexBuffer ().Size (),
                           max_primitives_count = 0,
                           max_count      = indices_count ? indices_count : vertices_count;
@@ -339,7 +339,7 @@ class BinMeshLibraryLoader
       
         //чтение идентификатора меша
         
-      size_t id_length;
+      uint32_t id_length;
       
       file_read (input_file, &id_length, sizeof (id_length));
       
@@ -351,7 +351,7 @@ class BinMeshLibraryLoader
 
         //чтение имени меша
 
-      size_t name_length;
+      uint32_t name_length;
 
       file_read (input_file, &name_length, sizeof (name_length));
 
@@ -369,7 +369,7 @@ class BinMeshLibraryLoader
 
         //чтение индексного буфера
 
-      size_t has_index_buffer;
+      uint32_t has_index_buffer;
 
       file_read (input_file, &has_index_buffer, sizeof (has_index_buffer));
 
@@ -380,14 +380,14 @@ class BinMeshLibraryLoader
         if (has_index_buffer > 1)
           throw xtl::format_operation_exception (METHOD_NAME, "Mesh index buffers count more than 1");
 
-        size_t ib_id;
+        uint32_t ib_id;
 
         file_read (input_file, &ib_id, sizeof (ib_id));
 
         IndexBufferMap::iterator ib_iter = index_buffers.find (ib_id);
 
         if (ib_iter == index_buffers.end ())
-          throw xtl::make_argument_exception (METHOD_NAME, "index_buffer", ib_id, "Index buffer not found");
+          throw xtl::make_argument_exception (METHOD_NAME, "index_buffer", (size_t)ib_id, "Index buffer not found");
 
         ib = &ib_iter->second;
       }
@@ -397,31 +397,31 @@ class BinMeshLibraryLoader
 
         //поиск вершинных буферов
 
-      size_t vertex_buffers_count;
+      uint32_t vertex_buffers_count;
 
       file_read (input_file, &vertex_buffers_count, sizeof (vertex_buffers_count));
 
-      for (size_t i = 0; i < vertex_buffers_count; i++)
+      for (uint32_t i = 0; i < vertex_buffers_count; i++)
       {
-        size_t vb_id;
+        uint32_t vb_id;
 
         file_read (input_file, &vb_id, sizeof (vb_id));
 
         VertexBufferMap::iterator iter = vertex_buffers.find (vb_id);
 
         if (iter == vertex_buffers.end ())
-          throw xtl::make_argument_exception (METHOD_NAME, "vertex_buffers[i]", vb_id, "Vertex buffer not found");
+          throw xtl::make_argument_exception (METHOD_NAME, "vertex_buffers[i]", (size_t)vb_id, "Vertex buffer not found");
 
         mesh.Attach (iter->second);
       }
 
         //чтение примитивов
 
-      size_t primitives_count;
+      uint32_t primitives_count;
 
       file_read (input_file, &primitives_count, sizeof (primitives_count));
 
-      for (size_t i = 0; i < primitives_count; i++)
+      for (uint32_t i = 0; i < primitives_count; i++)
         ReadPrimitive (mesh);
 
         //присоединение меша к модели
@@ -432,11 +432,11 @@ class BinMeshLibraryLoader
       //чтение мешей
     void ReadMeshes ()
     {
-      size_t meshes_count;
+      uint32_t meshes_count;
 
       file_read (input_file, &meshes_count, sizeof (meshes_count));
 
-      for (size_t i = 0; i < meshes_count; i++)
+      for (uint32_t i = 0; i < meshes_count; i++)
         ReadMesh ();
     }
 
@@ -479,10 +479,10 @@ class BinMeshLibraryLoader
     }
 
   private:
-    typedef stl::hash_map<size_t, VertexStream>       VertexStreamMap;
-    typedef stl::hash_map<size_t, VertexWeightStream> VertexWeightStreamMap;
-    typedef stl::hash_map<size_t, VertexBuffer>       VertexBufferMap;
-    typedef stl::hash_map<size_t, IndexBuffer>        IndexBufferMap;
+    typedef stl::hash_map<uint32_t, VertexStream>       VertexStreamMap;
+    typedef stl::hash_map<uint32_t, VertexWeightStream> VertexWeightStreamMap;
+    typedef stl::hash_map<uint32_t, VertexBuffer>       VertexBufferMap;
+    typedef stl::hash_map<uint32_t, IndexBuffer>        IndexBufferMap;
 
   private:
     InputFile             input_file;     //исходный файл
